LLM_server/requirements.txt
Django
djangorestframework
python-dotenv
langchain-core 
langchain-community 
langchain-openai 

openai 
tiktoken 
ormsgpack
django-redis
pytz
uvloop>=0.17.0
httpx
six>=1.16.0

asgiref>=3.6.0
google-api-python-client
confluent-kafka
# Fix protobuf version conflict
protobuf>=4.21.0,<6.0.0
grpcio>=1.50.0
grpcio-health-checking>=1.50.0
pydub
# 🚀 현업 레벨 최적화 패키지
numpy>=1.24.0
asyncio-throttle
uvicorn[standard]
channels
langchain_groq
torch
transformers
accelerate
langchain-ollama